[Feat] 컴포넌트 파일구조 변경 및 분리
