+++ daisso/src/component/header/styles.jsx
@@ -1,21 +1,4 @@
-import React from "react";
 import styled from "styled-components";
-
-function Header() {
-  return (
-    <HeaderDiv>
-      <HeaderLogo>
-        <img src="../daisso.jpeg" width="75" height="75" />
-      </HeaderLogo>
-      <HeaderRight>
-        <div>로그인</div>
-        <div>회원가입</div>
-      </HeaderRight>
-    </HeaderDiv>
-  );
-}
-
-export default Header;
 
 const HeaderDiv = styled.div`
   border-bottom: 1.5px solid #da3731;
@@ -39,3 +22,5 @@ const HeaderRight = styled.div`
   align-items: center;
   color: #464646;
 `;
+
+export { HeaderDiv, HeaderLogo, HeaderRight };
