+++ daisso/src/component/layout/Layout.jsx
@@ -1,7 +1,7 @@
 import React from "react";
 import styled from "styled-components";
-import Header from "./Header";
-import Footer from "../component/Footer";
+import Header from "../header/Header";
+import Footer from "../footer/Footer";
 
 const Layout = ({ children }) => {
   return (
