+++ daisso/src/component/layout/styles.jsx
@@ -1,0 +1,7 @@
+import styled from "styled-components";
+
+const LayoutLimit = styled.div`
+  min-width: 1000px;
+`;
+
+export { LayoutLimit };
