+++ daisso/src/component/login/styles.jsx
@@ -1,0 +1,3 @@
+import styled from "styled-components";
+
+export {};
