+++ daisso/src/pages/MainPage.jsx
@@ -1,7 +1,7 @@
 import React from "react";
-import Layout from "../component/Layout";
-import Main from "../component/Main";
-import Post from "../component/Post";
+import Layout from "../component/layout/Layout";
+import Main from "../component/main/Main";
+import Post from "../component/post/Post";
 
 function MainPage() {
   return (
